test: separates slow tests into separate files
